requirements.txt
spotipy
psycopg2
boto3
python-dotenv
pandas